MSI: slight refactoring
