+++ app/code/Magento/Inventory/Model/StockSourceLink/Command/GetAssignedSourcesForStock.php
@@ -3,22 +3,23 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\Inventory\Model;
+namespace Magento\Inventory\Model\StockSourceLink\Command;
 
 use Magento\Framework\Api\SearchCriteria\CollectionProcessorInterface;
 use Magento\Framework\Api\SearchCriteriaBuilder;
-use Magento\Framework\Exception\CouldNotDeleteException;
 use Magento\Framework\Exception\InputException;
-use Magento\Inventory\Model\ResourceModel\StockSourceLink as StockSourceLinkResourceModel;
+use Magento\Framework\Exception\LocalizedException;
 use Magento\Inventory\Model\ResourceModel\StockSourceLink\Collection;
 use Magento\Inventory\Model\ResourceModel\StockSourceLink\CollectionFactory;
-use Magento\InventoryApi\Api\UnassignSourceFromStockInterface;
+use Magento\InventoryApi\Api\Data\SourceInterface;
+use Magento\InventoryApi\Api\GetAssignedSourcesForStockInterface;
+use Magento\InventoryApi\Api\SourceRepositoryInterface;
 use Psr\Log\LoggerInterface;
 
 /**
  * @inheritdoc
  */
-class UnassignSourceFromStock implements UnassignSourceFromStockInterface
+class GetAssignedSourcesForStock implements GetAssignedSourcesForStockInterface
 {
     /**
      * @var CollectionProcessorInterface
@@ -36,9 +37,9 @@ class UnassignSourceFromStock implements UnassignSourceFromStockInterface
     private $searchCriteriaBuilder;
 
     /**
-     * @var StockSourceLinkResourceModel
+     * @var SourceRepositoryInterface
      */
-    private $stockSourceLinkResource;
+    private $sourceRepository;
 
     /**
      * @var LoggerInterface
@@ -49,48 +50,59 @@ class UnassignSourceFromStock implements UnassignSourceFromStockInterface
      * @param CollectionProcessorInterface $collectionProcessor
      * @param CollectionFactory $stockLinkCollectionFactory
      * @param SearchCriteriaBuilder $searchCriteriaBuilder
-     * @param StockSourceLinkResourceModel $stockSourceLinkResource
+     * @param SourceRepositoryInterface $sourceRepository
      * @param LoggerInterface $logger
      */
     public function __construct(
         CollectionProcessorInterface $collectionProcessor,
         CollectionFactory $stockLinkCollectionFactory,
         SearchCriteriaBuilder $searchCriteriaBuilder,
-        StockSourceLinkResourceModel $stockSourceLinkResource,
+        SourceRepositoryInterface $sourceRepository,
         LoggerInterface $logger
     ) {
         $this->collectionProcessor = $collectionProcessor;
         $this->stockLinkCollectionFactory = $stockLinkCollectionFactory;
         $this->searchCriteriaBuilder = $searchCriteriaBuilder;
-        $this->stockSourceLinkResource = $stockSourceLinkResource;
+        $this->sourceRepository = $sourceRepository;
         $this->logger = $logger;
     }
 
     /**
      * @inheritdoc
      */
-    public function execute($stockId, $sourceId)
+    public function execute($stockId)
     {
-        if (!is_numeric($stockId) || !is_numeric($sourceId)) {
+        if (!is_numeric($stockId)) {
             throw new InputException(__('Input data is invalid'));
         }
+        try {
+            $sourceIds = $this->getAssignedSourceIds($stockId);
 
+            $searchCriteria = $this->searchCriteriaBuilder
+                ->addFilter(SourceInterface::SOURCE_ID, $sourceIds, 'in')
+                ->create();
+            $searchResult = $this->sourceRepository->getList($searchCriteria);
+            return $searchResult->getItems();
+        } catch (\Exception $e) {
+            $this->logger->error($e->getMessage());
+            throw new LocalizedException(__('Could not load Sources for Stock'), $e);
+        }
+    }
+
+    /**
+     * @param int $stockId
+     * @return array
+     */
+    private function getAssignedSourceIds($stockId)
+    {
+        // TODO: replace on direct SQL query
         $searchCriteria = $this->searchCriteriaBuilder
             ->addFilter(StockSourceLink::STOCK_ID, (int)$stockId)
-            ->addFilter(StockSourceLink::SOURCE_ID, $sourceId)
             ->create();
-
         /** @var Collection $collection */
         $collection = $this->stockLinkCollectionFactory->create();
         $this->collectionProcessor->process($searchCriteria, $collection);
-        $items = $collection->getItems();
-
-        try {
-            $stockSourceLink = reset($items);
-            $this->stockSourceLinkResource->delete($stockSourceLink);
-        } catch (\Exception $e) {
-            $this->logger->error($e->getMessage());
-            throw new CouldNotDeleteException(__('Could not delete Source Stock Link'), $e);
-        }
+        $data = $collection->getData();
+        return $data ? array_column($data, 'source_id') : [];
     }
 }
