+++ app/code/Magento/Inventory/Model/StockSourceLink/Command/UnassignSourceFromStock.php
@@ -3,23 +3,22 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\Inventory\Model;
+namespace Magento\Inventory\Model\StockSourceLink\Command;
 
 use Magento\Framework\Api\SearchCriteria\CollectionProcessorInterface;
 use Magento\Framework\Api\SearchCriteriaBuilder;
+use Magento\Framework\Exception\CouldNotDeleteException;
 use Magento\Framework\Exception\InputException;
-use Magento\Framework\Exception\LocalizedException;
+use Magento\Inventory\Model\ResourceModel\StockSourceLink as StockSourceLinkResourceModel;
 use Magento\Inventory\Model\ResourceModel\StockSourceLink\Collection;
 use Magento\Inventory\Model\ResourceModel\StockSourceLink\CollectionFactory;
-use Magento\InventoryApi\Api\Data\SourceInterface;
-use Magento\InventoryApi\Api\GetAssignedSourcesForStockInterface;
-use Magento\InventoryApi\Api\SourceRepositoryInterface;
+use Magento\InventoryApi\Api\UnassignSourceFromStockInterface;
 use Psr\Log\LoggerInterface;
 
 /**
  * @inheritdoc
  */
-class GetAssignedSourcesForStock implements GetAssignedSourcesForStockInterface
+class UnassignSourceFromStock implements UnassignSourceFromStockInterface
 {
     /**
      * @var CollectionProcessorInterface
@@ -37,9 +36,9 @@ class GetAssignedSourcesForStock implements GetAssignedSourcesForStockInterface
     private $searchCriteriaBuilder;
 
     /**
-     * @var SourceRepositoryInterface
+     * @var StockSourceLinkResourceModel
      */
-    private $sourceRepository;
+    private $stockSourceLinkResource;
 
     /**
      * @var LoggerInterface
@@ -50,59 +49,48 @@ class GetAssignedSourcesForStock implements GetAssignedSourcesForStockInterface
      * @param CollectionProcessorInterface $collectionProcessor
      * @param CollectionFactory $stockLinkCollectionFactory
      * @param SearchCriteriaBuilder $searchCriteriaBuilder
-     * @param SourceRepositoryInterface $sourceRepository
+     * @param StockSourceLinkResourceModel $stockSourceLinkResource
      * @param LoggerInterface $logger
      */
     public function __construct(
         CollectionProcessorInterface $collectionProcessor,
         CollectionFactory $stockLinkCollectionFactory,
         SearchCriteriaBuilder $searchCriteriaBuilder,
-        SourceRepositoryInterface $sourceRepository,
+        StockSourceLinkResourceModel $stockSourceLinkResource,
         LoggerInterface $logger
     ) {
         $this->collectionProcessor = $collectionProcessor;
         $this->stockLinkCollectionFactory = $stockLinkCollectionFactory;
         $this->searchCriteriaBuilder = $searchCriteriaBuilder;
-        $this->sourceRepository = $sourceRepository;
+        $this->stockSourceLinkResource = $stockSourceLinkResource;
         $this->logger = $logger;
     }
 
     /**
      * @inheritdoc
      */
-    public function execute($stockId)
+    public function execute($sourceId, $stockId)
     {
-        if (!is_numeric($stockId)) {
+        if (!is_numeric($sourceId) || !is_numeric($stockId)) {
             throw new InputException(__('Input data is invalid'));
         }
-        try {
-            $sourceIds = $this->getAssignedSourceIds($stockId);
 
-            $searchCriteria = $this->searchCriteriaBuilder
-                ->addFilter(SourceInterface::SOURCE_ID, $sourceIds, 'in')
-                ->create();
-            $searchResult = $this->sourceRepository->getList($searchCriteria);
-            return $searchResult->getItems();
-        } catch (\Exception $e) {
-            $this->logger->error($e->getMessage());
-            throw new LocalizedException(__('Could not load Sources for Stock'), $e);
-        }
-    }
-
-    /**
-     * @param int $stockId
-     * @return array
-     */
-    private function getAssignedSourceIds($stockId)
-    {
-        // TODO: replace on direct SQL query
         $searchCriteria = $this->searchCriteriaBuilder
             ->addFilter(StockSourceLink::STOCK_ID, (int)$stockId)
+            ->addFilter(StockSourceLink::SOURCE_ID, $sourceId)
             ->create();
+
         /** @var Collection $collection */
         $collection = $this->stockLinkCollectionFactory->create();
         $this->collectionProcessor->process($searchCriteria, $collection);
-        $data = $collection->getData();
-        return $data ? array_column($data, 'source_id') : [];
+        $items = $collection->getItems();
+
+        try {
+            $stockSourceLink = reset($items);
+            $this->stockSourceLinkResource->delete($stockSourceLink);
+        } catch (\Exception $e) {
+            $this->logger->error($e->getMessage());
+            throw new CouldNotDeleteException(__('Could not delete Source Stock Link'), $e);
+        }
     }
 }
